Fixed some warnings in Merlin and core

+++ sst/core/cfgoutput/pythonConfigOutput.cc
@@ -163,7 +163,7 @@ void PythonConfigGraphOutput::generate(const Config* cfg,
 bool PythonConfigGraphOutput::isMultiLine(const std::string check) const {
 	bool isMultiLine = false;
 
-	for(int i = 0; i < check.size(); i++) {
+	for(size_t i = 0; i < check.size(); i++) {
 		if(check.at(i) == '\n' || check.at(i) == '\r' || check.at(i) == '\f') {
 			isMultiLine = true;
 			break;
